feat(tactic): introduce a monad for interacting with type_context_old.
The idea is to investigate exposing some of the constructs which Lean
uses to make type_context_old to enable the user to create
more powerful tactics.

Some examples can be seen in the `tests/lean/tco.lean`:
There is a reimplementation of `intro` and `match_pattern` using the new monad.