fix(src/library/documentation.cpp): don't add lean suffix if code fence has an info string.
only add an info string/language to a code fence like:
```
def this : unit := ()
```

Code fences like:
```lean
def this : unit := ()
```

``` lean
def this : unit := ()
```
already have one, specified and a code fence like ```#check ()``` cannot have one.
Previously this could would turn ``` lean
				     def this : unit := ()
				 ```
into ```lean lean
        def this : unit := ()
     ```

and ```lean
       def this : unit := ()
    ```

    ```lean
       def this : unit := ()
    ```
into: ```lean
         def this : unit := ()
      ```lean

Since it only recognized the start of a code block if the 4th
character was whitespace. It would try and add an info string
to the end of the code fence, beleiving it to be the start of
a new block.