Fix `well_founded_tactics.default` on already solved goals
This fix is an adaptation from `well_founded_tactics.default_dec_tac'` in
mathlib, written by Scott Morrison.
The call to `unfold_sizeof` could happen to discharge the goal, and
`cancel_nat_add_lt` assumes that there are remaining goals. A call to `done`
checks for this condition, and finishes execution in that case.