fix(test/non_meta_rec_fn): fix flaky test by not checking stdout (#104)
This tests checks soundness of Lean when defining a generally recursive
function in a tactic block, by verifying this gives an error.  In CI there were
some failures because the error message given by `non_meta_rec_fn` did not have
a consistent column number (but was otherwise consistent). To make the tests
less flaky, moving it to the fail tests should check that the definition is
invalid, while not being as strict about the precise error message.