fix(library/type_context): inconsistent unification of match-bas… (#141)
The issue was that `nat.add` didn't get unified with `λ a b, nat.add a b`,
because `nat.add` is defined using matches. If such a definition is applied to
a term, only when the term matches one of the match cases the whole is
unfolded.  In contrast, if there is no application, the constant `nat.add` is
happily unfolded all the way. Thus the elaborator decides unification fails.

There are basically two solutions: unfolding more (i.e. also unfold `nat.add a
b`) or unfolding less (i.e. don't unfold `nat.add`). Since unfolding `nat.add`
would expose implementation details, namely the `nat.add._main` function,
choosing to unfold less is better.

But this leads to another issue: now `λ <a, b>, a` does not unify with
`λ <a, b>, a` if they are defined separately. This is not a big issue since
`example : let c := 0 in (λ <a, b>, a + c) = (λ <a, b>, a + c) := rfl`
didn't work anyway. Hence the changed tests.