fix(library/init): remove `if_simp_congr` for more powerful `if`… (#159)
The removed lemma caused `simp` to always use `decidable_of_decidable_iff`
to construct a `decidable` instance for the condition of an `if` expression.
That meant that `if c then t else e = if c' then t else e` often did not get
solved by `simp`, getting stuck on `dec_c = decidable_of_decidable_of_iff ...`
There is another `congr` lemma, `if_congr`, defined just before
`if_simp_congr`, which does allow the `decidable` instances to differ.
Because of the order of the declarations, `if_simp_congr` takes priority.

It seems that `if_simp_congr` was introduced to allow progress rewriting
`if c then t else e` to `if c' then t else e` without requiring a
`decidable c'` instance. However, this progress should never be needed to
close a goal: either the `if` expression is rewritten by another `simp` lemma
carrying its own `decidable` instances, or it is not rewritten, so the
`decidable` instance already exists in the goal. The only breakage will be to
non-terminal `simp`s, which are an anti-pattern anyway.  The removed lemmas are
not used either in another part of the core library or mathlib.

The only errors occuring when compiling mathlib with these changes are goals
being closed too soon because `simp` has become more powerful.