doc(meta): fix some documentation todos
I removed some todos that were done or not necessary.
I resolved some of the todos be writing the docs.
Some of the [todo]ed declarations remain opaque to me, so
I've either removed them or added a link to a comment in the C++.
I don't think this is a regression because there are plenty of other undocumented
declarations that would need to be todoed to remain consistent.
The todos were originally there just as a personal note to attempt to
understand how they work, but they were left in during the doc PR.