fix(library/init/meta/interactive): removing for `simp [← local_const ...]`
This fixes issue #193.

The code to erase a lemma `h` from the simp set for calls of the form
`simp [←h]` matched on `h = expr.const _ _`. It should also match when `h` is a
`local_const`. I think this issue was not caught by the test because the the
arguments to `simp` were quoted incorrectly, so I've changed the test to use
the tactic quoter instead.