feat(src/frontends): order notation by priority in pretty-printer
If you have a notation for a special case, e.g. `finset.sum finset.univ`, and a
notation for a general case, e.g. `finset.sum`, you can now give the
special-cased notation a higher priority to ensure it is used by the
pretty-printer when applicable.

When deciding how to pretty-print an expression, Lean looks up the notations
associated to the head of the expression. It tries them one-by-one until it
finds a notation that matches the expression. The order in which they are tried
used to be the definition order. This PR changes this ordering to have
higher-priority notations first.

Although `notation_entry.priority()` is also used by the parser, it doesn't
seem to be set at all in mathlib. I expect that there will be no conflicts
between the usage of `priority` in the parser and in the pretty-printer.