fix(init/core): lower precedence of unary `-`
This commit moves the precedence of unary `-` below that of `^`
so that `-a^n` parses as `-(a^n)` and not (as formerly) `(-a)^n`.

It would be reasonable to move `-` below `*` as well (so that
`-a*b` parses as `-(a*b)`) but this is surprisingly disruptive
to mathlib, because `∩` and `⊓` have the same precedence as `*`
and `-` is also used for set complement. So, at least for now,
we leave the relative precedence of `-` and `*` unchanged.