feat(library/init/meta/tactic): informative tactic combinators (#212)
Many tactic combinators currently only work with `tactic unit`s. This PR introduces variants of these combinators which work with general `tactic α`s, as suggested in #206.

This is a choose-your-own-PR: It contains two commits corresponding to two different designs from #206. The first commit is more conservative: it leaves the existing combinators alone and implements the new ones with primed names (`all_goals'` etc). This leads to no breakage in core and very minimal breakage in mathlib. The second commit switches the names around so that the `unit`-valued combinators get the prime. This leads to mild breakage in core/mathlib, but is more consistent with the existing naming conventions (e.g. `mmap` vs `mmap'`).

I'd prefer that both commits get merged, but if you want just the first one, that's also fine by me.

fixes #206