chore(github): speed up emscripten build (#225)
Enables `-j2` in the emscripten build.  This didn't work before because the `gmp` target doesn't work with `-j2` so it has to be built separately.  This change saves about 15 minutes and brings the emscripten build time down to the time taken by the debug builds.

I've also tried only building the `lean_js_wasm` and `lean_js_js` targets.  This would save another 14 minutes, but I'd like to keep building the standard library as a sanity check.

Enabling the tests would also be possible, but they take a very long time.  (45 minutes just for the tests.)  A few tests fail, namely the IO tests and the type context test.  They would need to be commented out or fixed.