feat(widget): fix component reconcile issue
Components were not reconciling properly because I was checking equality
with pointer arithmetic. Now it uses hashes of VM objects.
The problem is that I didn't implement a hash for each of the
vm_externals so I think it is likely that it will still fail if a vm
external is used.