feat(vm): vm objects may be hashed
This is needed as an optimisation for widgets. To quickly check equality
of a pair of vm objects.
At the moment each instance of vm_external
needs to implement its own `hash` with a fallback of zero.
A warning is in the comment for hash.
For the purpose that I wrote hash for this is the most useful failure mode