feat(vm): vm objects may be hashed (#262)
This is needed as an optimisation for widgets. As a quick check of equality
of a pair of vm objects.
At the moment each instance of vm_external needs to implement its own `hash` with a fallback of zero.
A warning of this fallback is in the comment for hash.
It will be prohibitively difficult to implement hashing on all the vm_external members and for the purpose I built it for I would rather have false equality than false non-equality.

It also just recomputes the hash each time rather than caching it, since it will be calculated for the vast minority of vm objects this seems like the right thing to do.

Any comments on how to improve the vm hashing function (eg to remove the magic numbers) would be appreciated.

Co-authored-by: Ed Ayers <EdAyers@users.noreply.github.com>