fix(eformat): fix a potential address bug on coe
The problem is that `pp_hide_coercion` with more than 4 args rewrites the expression and then passes this to pp_child. But now you can't compute the address properly because it's not the true expression being pretty printed.
The ideal solution would be to rewrite `pp_child` to support adjusted function addresses.
But for now it just stops trying to calculate addresses on subterms which use function coes with more than 4 args.
This is quite a rare case anyway.