feat(eformat): add support for structured formatting (#276)
- Introduce a new datatype `tagged_format` for making
  formats where certain nodes can be tagged with   information
- Make the pretty printer templated so that it can support both `format`
   and the new tagged_format
- Add a tactic_state.pp_tagged : tactic_state → expr → eformat

In an ideal world. One would implement tags directly on to `format` however because format is implemented as a sexpr and because the widget module needs to be able to perform induction on the format object, I concluded that adding tags to `format` would amount to an overhaul of `format` which seems too big a change.