fix(vm_override): indexes are not preserved across instances
When VM_override was serialising a declaration, it was storing the
vm decl index of the override.
However there is no guarantee that the same decl will be given the
same index in different sessions, since indexes are not stored in the
olean.
Therefore when vm_code_modification is run it must serialise the
name of the override decl instead of the index.