Fix guards  to make it possible to build for BSD systems
These guards currently assumes that if a system isn't GNU/Linux, macOS or
Emscripten than it's necessarily Windows, which is of course false as it misses
a lot of other platforms.  This patch allows building for FreeBSD (and likely
for other BSD systems) by adopting the opposite logic: it distinguish between
Windows and non-Windows platforms, without having to list all of them
explicitly.