fix(shell/server): do not cancel info queries, etc. (#308)
If a second info request was sent to the server while the first one was still running, we used to cancel the first one.  This was always only a precaution to prevent high CPU usage.

Now that we're sending many info requests from various parts of the vscode extension to the server at the same time, this behavior is annoying because you get "interrupted" errors for no good reason.