fix(widget): Set handler_ids to not be set by a global counter.
If two widget events are fired together (as often happens for hovers).
Then the event handler indices for the re-rendered component would update
and make the second widget_update fail when it should have been processed.
A workaround is to just set the index for each component.
Although this could potentially introduce a bug later see comment for
`render_event` in widget.cpp.