perf(library/type_context): cache type-class searches w/o mvars (#332)
This PR adds caching of intermediate results during type-class synthesis, under the condition that the subgoals have no temporary metavariables.  For example, a subgoal `has_add α` is cached, but `has_coe_t α ?m` is not.

It is theoretically possible that some type-class searches fail now, because we no longer try all possible instances (and if this actually matters for another instance to apply).  But it's a bug anyhow if you have two instances that are not definitionally equal.