feat(interactive_expr): collect locals with the same type and local filtering (#346)
This gives a new 'filter' dropdown on the tactic state which can be used to filter the locals.
I've just added three filters but hopefully it is clear how to extend it. In order to interact with the dropdown you need this PR: https://github.com/leanprover/vscode-lean/pull/182

The problem is that since all widget state is local, you can't persist the filter choice as the user clicks around. So in order for the filter state to persist some sort of global widget persistence model needs to be implemented in Lean C++. That will require some significant work so I've just not implemented it for now.

Also if there are multiple consecutive locals with the same type then they will be collected in to a single row of the tactic state context display.