feat(widget): implement widget effects
save_widget now has the type `pos -> component tactic_state
widget.effects`.
This means that now the widget-writer can cause widget.effects
to issue a stream of commands to vscode to run side effects. Things like
highlighting a particular region of code or revealing a location in the
editor.

The next task is to add 'go to definition' for the default tactic
widget.