fix(pp_tagged): fix address issue
Some pp_tagged were using a different convention for which way round the
addresses go, meaning the addresses came out wonky. The convention
from now until the end of time is that the _head_ of the address
is the _most local_. However when pp_tagged is called in lean,
_this direction is reversed_, so now the head coordinate is the most global in lean.

There is an additional issue where (int.of_nat = coe) will interpret
pping coe as without the implicit arguments. I'm not sure whether that
would count as a fix so I left it as it is for now.