fix(frontends/lean/elaborator): honor as_is attribute for functions  (#399)
This is my attempt to fix #395. This is my first time digging into the Lean internals, so I apologize if my approach is a bad way to address the issue. I'm also unsure whether the two lines should be added in front of the if/else block checking for `@` mode or after it, but I wasn't able to get the elaborator to produce an expression containing an `@` that could be put into an `as_is`, so it might actually not matter.