refactor(init/data/nat,int): less choice (#446)
This PR removes usage of the axiom of choice from the basic nat and int lemmas. It also removes `private` from the internal lemmas about int (it is a pain to prove properties about `int.sub_nat_nat` when the library is locked up), and moves the `decidable.*` order theorems from mathlib. (While I think we should move order theory out of core, that is a separate PR, and if and when that happens these theorems can move back to mathlib. But for now they should be treated as dependents of the `nat.lemmas` file, and can migrate with it.)