chore(init/meta/interactive): use `guard_hyp h : t` (#445)
instead of `guard_hyp h := t`. This affects many files but it is a simple
find and replace.

# Motivation
If `h` has type `t` then we almost universally write that `h : t` in lean. Saying `h := t` looks more appropriate for checking the value of a let binding in the context rather than the type.