fix(init/meta/converter/interactive): Do not ignore errors and tactic state changes in `for` and `find` (#482)
An example of the type of code that used to work and now fails:
```lean
import algebra.ring

example {a b c : ℕ} : (a * b) * c = (a * b) * c :=  begin
  conv {
    for (_ * _) [1] {
      rw this_isnt_a_lemma,  -- no error message?
    },
  },
end

example {a b c : ℕ} : (a * b) * c = (a * b) * c :=  begin
  conv {
    find (_ * _) {
      rw this_isnt_a_lemma,  -- no error message?
    },  -- error: pattern not found (not true!)
  },
end
```
Without this PR, users like me get no information on why their tactic failed, which is a very frustrating experience.

The problem here is that `ext_simplify_core` ignores all failures in its `pre` and `post` arguments - failures are treated as "there was nothing to simplify here, move on to the next one". In `find` and `for`, we need to distinguish the cases of failing due to the pattern not matching and failuring due to an error in the user-provided tactic.

This PR adds three new tactic primitives, `tactic.capture`, `tactic.unwrap`, and `tactic.resume`, the first two of which are inspired by the python functions [from the `trio-outcome` package](https://outcome.readthedocs.io/en/latest/api.html) with the same name. `capture` can be used to run the user-provided tactic and keep track of any errors that occured. `unwrap` is then used to re-throw the error once `ext_simplify_core` is done.

It turns out that not only does `ext_simplify_core` ignore all failures, but it resets the tactic state after it completes - which clears populated metavariables. To preserve the tactic state, we need `resume`.

---

A replacement for #476, since that PR does not have CI runs