feat(shell/server): add quiet / never_use_olean arguments
The quiet argument prevent parse messages from being printed on stdout. This allows to subclass server:: to create a tactic mode server with a clean stdout.
The never_use_olean argument enforces the full parsing of all .lean files. This is useful to build snapshots for every lean files. These snapshots can then be used to easily parse arbitrary tactics "as-if" we were inside a declaration of interest.