feat(print_cmd.cpp): print universe declarations
This prints the prefix `{u}` in
```lean
def {u} id : Π {α : Sort u}, α → α :=
λ {α : Sort u} (a : α), a
```
when `pp.universes` is turned on. This is much easier than hunting down
all references to universes in the type and value of the theorem to
figure out how many universe arguments there are (and what order they
are in).