fix(checker/text_import): make leanchecker agree with external tc
The change to `text_import.cpp` checks the line for parse errors after
parsing, so we don't accidentally let malformed lines pass.

The change to the kernel removes the check for anonymous names in
lambda/let/pi. Since these are used only for pretty printing it should
not affect soundness. An alternative approach would be to automatically
generate a fresh name when the given name is anonymous.