refactor(init/core): backport remove reserve notation (#599)
This removes all uses of `reserve notation` in lean core (mostly in `init.core`), which is not supported in lean 4. The main intent here is to make sure that later syntax declarations come with the precedence information attached directly, instead of being pre-declared.

Additionally, lean was modified to allow writing:
```lean
reserve infix `foo`:3
infix `foo`:3 := bar
```
Previously this was an error, you are *required* to omit the `:3` in the second declaration. Now you are allowed to write it but it still checks that it is `3` (so ``infix `foo`:4`` would be an error).

Finally, `::` is now a built in token instead of being defined in the first few lines of `init.core`. This is needed because without the `reserve notation` declaring it, it's not a token at all, even though `::` is used in structure syntax, e.g. `structure foo := mk :: (a : Prop)`. Since there are some structures that have to be defined before `list.cons` and its notation, this causes a parse error.