refactor(init/core): backport remove unification hints (#598)
This one is a no-brainer. It is well known that lean 3 practically does not use unification hints at all. There is exactly one use of the `@[unify]` attribute in lean 3 / mathlib, so let's just remove it. Lean 4 does support unification hints but the syntax is different enough and there is only one example so alignment isn't really worth it.

This might have unforeseen fallout in mathlib. We will see how much we were actually relying on this unification hint.