feat(parser): AST tracking and export (#608)
This PR implements AST tracking across the whole parser. That means that every parser command creates an AST node, and the result can currently be exported via the `--ast` command line argument, which produces a `foo.ast.json` file for every `foo.lean` file that is processed (note that files that already have oleans will not be processed and so will not have ast files generated).

The basic schema of the exported JSON file (see `ast_exporter.cpp`) looks like this:
```ts
{
  ast: [null | {
    kind: string,
    value?: name,
    start: [int, int],
    end?: [int, int],
    children?: [int],
    pexpr?: int,
    expr?: int,
  }],
  file: int,
  level: [
    "0" |
    {suc: int} |
    {max: [int, int]} |
    {imax: [int, int]} |
    {param: int} |
    {mvar: int}
  ],
  expr: [
    {var: int} |
    {sort: int} |
    {mvar: {name: name, pp: name, type: int}} |
    {local: {name: name, pp: name, bi: int, type: int}} |
    {const: [name, [int]]} |
    {Pi: {name: name, bi: int, dom: int, body: int}} |
    {lam: {name: name, bi: int, dom: int, body: int}} |
    {app: [int, int]} |
    {let: {name: name, type: int, value: int, body: int}} |
    ... // one variant per macro
  ]
}
```
The https://github.com/dselsam/mathport repo contains an .ast.json parser developed in tandem with this PR, which can be used as a more precise specification for the files, at least until a proper specification is written.

The behavior of lean with the `--ast` option disabled should be otherwise unchanged, although memory usage might possibly be a bit higher. It is possible to compile the AST stuff out if necessary, or disable the collection at runtime, but this version does not support that. Future extensions may add direct access to the AST data from within lean.