feat(ast,scanner): add end-pos support  (#612)
This adds end position support to the scanner and therefore the AST nodes. There's actually "two different" proposals in this PR for how this is implemented outside the scanner (in the two separate commits); the first is very light and definitely works, but is only implemented for `tactic`s, `<|>` and `;` (my usecase in Alectryon).

The second commit hijacks `set_ast_pexpr` to also set the AST's position `m_end`, but I'm not sure in what variety of ways this method is used and so it may give (and as far as I can see, somewhat does) give some false results, but it's at least correct in the above 3 settings.

I also modified `check_break_before` to use this new end-pos; this means that, for example, something like this:

```lean
example : 0 = 1 :=
begin
  symmetry           ,
  sorry

  --some comment
end
```

if you have the cursor anywhere between the `symmetry` and the comma, the tactic state will show the state <after> the symmetry, and similarly after the `sorry` token ends (although the "goals accomplished!" text can sometimes not show up). I think that I should do something similar for `check_break_at_pos`, but I left it for now as I don't wanna accidentally break too much.