fix(init/funext): make instance local and not private (#611)
The declaration `fun_setoid` was previously a private but global instance.

Initially I only made the instance local (I expect it is not used in mathlib anywhere). However, it's used in a test, so I made the declarations in this file non-private (also, because I hate private declarations).

I also moved the declarations `fun_setoid`, `extfun`, `fun_to_extfun` and `extfun_app` into the `function` namespace, and added some docstrings.