feat(tactic_log.cpp): log tactic invocations in AST (#614)
This augments the `--ast` option to generate files with tactic invocation data. The way it works is that `tactic.istep` calls the `with_ast` meta constant, which will log the tactic state before and after each tactic invocation, in addition to the AST node of the tactic itself. Tactic invocations are added to a file-global list, so they may end up reordered nondeterministically because of parallel proof processing, but the tactic calls for a single declaration should be in order as long as the `async` tactic is not used.

The type signature of `istep` is changed (to add an additional `ast` argument); this affects all tactic classes, including custom tactic classes, so this will break downstream libraries that use custom tactic classes (the fix is easy though, see `tests/lean/run/my_tac_class.lean`).