chore(library/scoped_ext): resolve `open` like `def` (#619)
This changes the namespace resolution for namespaces in `open foo` to prefer the current stack of namespaces over the root, to match name resolution in declarations. Like declarations, you can use `open _root_.foo` to override this and open the namespace named `foo` even if you are in a namespace `bar` and the namespace `bar.foo` exists.