fix(init/meta/interactive_base): Use `}}` in format (#650)
As [discussed on Zulip](https://leanprover.zulipchat.com/#narrow/stream/113488-general/topic/curly.20braces/near/261185802). This way, escaped braces are balanced. That is, `format!"a{{b}}c"` produces `"a{b}c"`, instead of `format!"a{{b}c"`. Unbalanced braces are bad for editors and formatters, so we should not encourage any syntax that requires unbalanced braces.