fix(library/util): unfold rec args as semireducible
This was causing an assertion violation. It is not clear what the best
fix is, since the code seems to be of two minds about what can be
unfolded, but since the ginductive has a whole error message explaining that
semireducible definitions are supposed to be unfolded, I went with that.