`simp` should test `ready_to_synthesize` before calling `mk_instance`
This ensures non-`out_param`s are not accidentally filled through unification
with the temporary metavariables. `out_param`s are still allowed to contain
metavariables that will be unified (see also #657).

I moved the `ready_to_synthesize` check from the elaborator to the type context
in order to access it from the simplifier.

Co-Authored-By: Gabriel Ebner <gebner@gebner.org>