feat(pp): add option and attribute to display type ascriptions for numerals (#691)
This provides a mechanism to cause the pretty printer to display type ascriptions when displaying numerals. They are shown either when the head of the type of the numeral has the `pp_numeral_type` attribute or when the `pp.numeral_types` option is set to `true`.

For example,
```lean
attribute [pp_numeral_type] fin
#check (2 : fin 17) + 3
-- (2 : fin 17) + (3 : fin 17) : fin 17

set_option pp.numeral_types true
#check (2 : fin 17) + 3
-- (2 : fin (17 : ℕ)) + (3 : fin (17 : ℕ)) : fin (17 : ℕ)
```

The pretty printer has special support for `nat` constructor expressions. For these, the ascription is displayed without additional notation, to signal that this is not a numeral in normal `bit0`/`bit1` form.
```lean
set_option pp.numeral_types true
#check nat.zero.succ.succ.succ
-- (3 : nat) : ℕ
```