feat(modifiers): `noncomputable!` to force noncomputability (#703)
This adds a user-visible feature: the `noncomputable!` modifier. Definitions with this will not have their computability checked, will be marked noncomputable when added to the environment, and will not be VM compiled. Noncomputability modifiers interact with the `noncomputable theory` command in the following way:

1. Without `noncomputable theory`
   1. No modifier: check is computable, run VM compiler
   2. `noncomputable`: check is noncomputable, do not VM compile, mark declaration as noncomputable
   3. `noncomputable!`: do not test computability, do not VM compile, and mark declaration as noncomputable
2. With `noncomputable theory`
   1. No modifier or `noncomputable`: if computable, run VM compiler; if noncomputable, mark declaration as noncomputable and do not VM compile
   2. `noncomputable!`: do not test computability, do not VM compile, and mark declaration as noncomputable

A `noncomputable! theory` command is explicitly not supported to avoid the possibility of making the Lean 4 porting effort any more difficult.

Behind the scenes, there is a little bit of cleanup, refactoring, and additional internal documentation.

Co-authored-by: Mario Carneiro <di.gama@gmail.com>