fix(noncomputable,definition_cmds): require non-Prop theorems to be noncomputable (#704)
The noncomputability checker is meant to determine what can and cannot be VM compiled, and since theorems do not get VM compiled, non-Prop theorems should be noncomputable. This is more restrictive than necessary, but at least for mathlib every theorem must be Prop-valued to pass the linter anyway.

The noncomputability error messages are also somewhat improved.