fix(noncomputable, definition_cmds): better error message for non-Prop theorems
The error message for non-Prop theorems gives misleading advice -- you
likely want to switch to using a `def` rather than adding
`noncomputable`. This change also prevents the error from appearing if
the type of the theorem contains `sorry`, since a user is likely in
the process of editing the theorem and the `def`/`noncomputable`
advice is unhelpful.