feat(frontends/lean): pretty-print coercions with type ascriptions
One of the major issues with roundtripping of expressions through the pretty-printer is that coercions are often not elaborated correctly, because their expected type could not be inferred. I propose to add a type ascription to the pretty-printing of coercions, so that e.g. `((0 : ℕ) : ℝ)` gets pretty-printed as `(↑(0 : ℕ) : ℝ)`. This should also make the expected type clearer for the human reader, so goals such as `↑x = ↑y` should become less mysterious. This PR does the same for `coe_fn`, but not `coe_sort` because the expected type of `coe_sort` is not hard to infer.

A drawback I foresee of this change is that things become more verbose: a goal with many coercions into complicated types might become more difficult to read at a glance. I've included an option `pp.ascript_coercions` to turn the new pretty-printing off.