fix(frontends/lean/parser_config): serialize notation names (#759)
Addendum to #754. The underlying issue in https://github.com/leanprover-community/lean/pull/754#issuecomment-1224998590 was that the `.olean` files did not serialize the notation names, so although everything works in a single-session `lean --make` call, if you try to do it in multiple passes the read-in `.olean` files will not have the notation names and will cause a conflict.

This changes the olean format, but AFAIR there is no olean version or anything to bump.