fix(frontends/lean/pp): correct binder links (#781)
Previously these used the notation definition (the RHS of the `:=`) rather than the matched expression. This usually doesn't work at all, since often the RHS is just `#0` and all the notation happens within the `scoped` block.

This affected a lot of mathlib notation; `Exists`/`infi`/`supr`/`filter.eventually`/`filter.frequently`/...

I don't know if this used to work and then we broke it, or if it's always been broken.